Laying ground for possible analysis variants
I outlined here possible structure for Zonation analyses. This order is based on the idea that we are especially focusing on connectivity and performance of the GPAN. Therefore hierarchical mask and planning units are included already in the beginning. In addition to these analyses also some test sets needs to be run.